Moved test app to a separate folder
