+++ test/app/src/App.svelte
@@ -63,13 +63,13 @@
 <script>
 // Import the router component
 // Normally, this would be import: `import Router from 'svelte-spa-router'`
-import Router from '../../Router.svelte'
+import Router from '../../../Router.svelte'
 // Import the "link" action and the methods to control history programmatically from the same module, as well as the location store
-import {link, push, pop, replace, location, querystring} from '../../Router.svelte'
+import {link, push, pop, replace, location, querystring} from '../../../Router.svelte'
 
 // Import the "active" action
 // Normally, this would be import: `import active from 'svelte-spa-router/active'`
-import active from '../../active'
+import active from '../../../active'
 
 // Import the list of routes
 import routes from './routes'
