+++ test/app/src/routes.js
@@ -1,6 +1,6 @@
 // Import the "wrap" function
 // Normally, this would be import: `import {wrap} from 'svelte-spa-router/wrap'`
-import {wrap} from '../../wrap.js'
+import {wrap} from '../../../wrap.js'
 
 // Components
 import Catalog from './routes/Catalog.svelte'
