+++ test/app/src/routes/Nested.svelte
@@ -6,7 +6,7 @@
 <script>
 // Import the router component
 // Normally, this would be import: `import Router from 'svelte-spa-router'`
-import Router from '../../../Router.svelte'
+import Router from '../../../../Router.svelte'
 
 // Routes
 import Wild from './Wild.svelte'
